pyproject.toml
python = "3.11"

[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "bioguard-ai"
version = "1.0.0"
description = "Privacy-First Health Ecosystem with Real-Time AR"
dependencies = [
    "streamlit>=1.28.0",
    "streamlit-webrtc>=0.47.0", 
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.0",
    "plotly>=5.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "ultralytics>=8.0.0",
    "scikit-image>=0.21.0",
    "chromadb>=0.4.0",
    "networkx>=3.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "pyotp>=2.9.0",
    "cryptography>=41.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "loguru>=0.7.0",
    "tqdm>=4.66.0"
]